Further tweaks

+++ src/TrafficLight.cpp
@@ -85,7 +85,7 @@ void TrafficLight::cycleThroughPhases()
         std::this_thread::sleep_for(std::chrono::milliseconds(1));
 
         // compute time difference to stop watch
-        long timeSinceLastUpdate = std::chrono::duration_cast<std::chrono::seconds>(std::chrono::system_clock::now() - lastUpdate).count();
+        double timeSinceLastUpdate = std::chrono::duration_cast<std::chrono::seconds>(std::chrono::system_clock::now() - lastUpdate).count();
 
         std::unique_lock<std::mutex> lck(_mtx);
         std::cout << "cycleDuration: " << cycleDuration << std::endl;
